# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/webhooks": {
      "post": {
        "description": "The signing secret is returned once, here, and never again. Targets must be HTTPS on a public address; the same check runs again before every delivery.",
        "operationId": "WebhookController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWebhookDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookCreatedDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "This wallet is not an operator.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Register a webhook subscription",
        "tags": [
          "webhooks"
        ]
      },
      "get": {
        "description": "Secrets are reported as fingerprints only. `exhaustedDeliveries` counting up is a receiver that has stopped answering.",
        "operationId": "WebhookController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookListDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "This wallet is not an operator.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Every subscription, with its delivery backlog",
        "tags": [
          "webhooks"
        ]
      }
    },
    "/webhooks/{id}": {
      "patch": {
        "operationId": "WebhookController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWebhookDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookSubscriptionDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "This wallet is not an operator.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Change a subscription’s target, events or status",
        "tags": [
          "webhooks"
        ]
      },
      "delete": {
        "description": "Removes the subscription and its delivery history. The events themselves stay.",
        "operationId": "WebhookController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "This wallet is not an operator.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a subscription",
        "tags": [
          "webhooks"
        ]
      }
    },
...
  },
  "info": {
    "title": "Rivora Protocol API",
...
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "A token from `POST /auth/verify`."
      },
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Partner API key. Requires the `score:read` scope."
      }
    },
    "schemas": {
...
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Requested 99999.00 exceeds available credit of 530.00."
          },
          "code": {
            "type": "string",
            "example": "exceeds_available",
            "description": "Machine-readable reason, for branching in the client."
          },
          "statusCode": {
            "type": "number",
            "example": 422
          }
        },
        "required": [
          "error",
          "code",
          "statusCode"
        ]
      },
...
      "CreateWebhookDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "example": "https://ops.example.com/hooks/rivora",
            "description": "Delivery target. HTTPS with a public address; validated again at every delivery."
          },
          "events": {
            "type": "array",
            "example": [
              "credit.draw.completed",
              "borrower.defaulted"
            ],
            "description": "Event types to receive, from the PRD §27 catalogue.",
            "items": {
              "type": "string",
              "enum": [
                "revenue.received",
                "revenue.settled",
                "risk.assessment.completed",
                "credit.limit.updated",
                "credit.draw.completed",
                "credit.repayment.completed",
                "borrower.watchlisted",
                "borrower.restricted",
                "borrower.defaulted",
                "vault.utilization.changed"
              ]
            }
          }
        },
        "required": [
          "url",
          "events"
        ]
      },
      "WebhookCreatedDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cmb1x…"
          },
          "url": {
            "type": "string",
            "example": "https://ops.example.com/hooks/rivora"
          },
          "events": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "revenue.received",
                "revenue.settled",
                "risk.assessment.completed",
                "credit.limit.updated",
                "credit.draw.completed",
                "credit.repayment.completed",
                "borrower.watchlisted",
                "borrower.restricted",
                "borrower.defaulted",
                "vault.utilization.changed"
              ]
            }
          },
          "secretFingerprint": {
            "type": "string",
            "example": "9f2c11ab",
            "description": "First bytes of the secret’s SHA-256 — enough to tell two secrets apart, never enough to sign."
          },
          "active": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "type": "string",
            "example": "2026-08-05T00:00:00.000Z"
          },
          "secret": {
            "type": "string",
            "example": "whsec_4f1c…",
            "description": "HMAC-SHA256 signing key. Shown once, at creation, and never again."
          }
        },
        "required": [
          "id",
          "url",
          "events",
          "secretFingerprint",
          "active",
          "createdAt",
          "secret"
        ]
      },
      "WebhookSubscriptionSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cmb1x…"
          },
          "url": {
            "type": "string",
            "example": "https://ops.example.com/hooks/rivora"
          },
          "events": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "revenue.received",
                "revenue.settled",
                "risk.assessment.completed",
                "credit.limit.updated",
                "credit.draw.completed",
                "credit.repayment.completed",
                "borrower.watchlisted",
                "borrower.restricted",
                "borrower.defaulted",
                "vault.utilization.changed"
              ]
            }
          },
          "secretFingerprint": {
            "type": "string",
            "example": "9f2c11ab",
            "description": "First bytes of the secret’s SHA-256 — enough to tell two secrets apart, never enough to sign."
          },
          "active": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "type": "string",
            "example": "2026-08-05T00:00:00.000Z"
          },
          "pendingDeliveries": {
            "type": "number",
            "example": 0,
            "description": "Deliveries waiting for the dispatcher."
          },
          "exhaustedDeliveries": {
            "type": "number",
            "example": 0,
            "description": "Deliveries dropped after every attempt failed."
          }
        },
        "required": [
          "id",
          "url",
          "events",
          "secretFingerprint",
          "active",
          "createdAt",
          "pendingDeliveries",
          "exhaustedDeliveries"
        ]
      },
      "WebhookListDto": {
        "type": "object",
        "properties": {
          "subscriptions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WebhookSubscriptionSummaryDto"
            }
          }
        },
        "required": [
          "subscriptions"
        ]
      },
      "UpdateWebhookDto": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "example": "https://ops.example.com/hooks/rivora"
          },
          "events": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "revenue.received",
                "revenue.settled",
                "risk.assessment.completed",
                "credit.limit.updated",
                "credit.draw.completed",
                "credit.repayment.completed",
                "borrower.watchlisted",
                "borrower.restricted",
                "borrower.defaulted",
                "vault.utilization.changed"
              ]
            }
          },
          "active": {
            "type": "boolean",
            "example": false,
            "description": "A disabled subscription receives nothing and its queue exhausts."
          }
        }
      },
      "WebhookSubscriptionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cmb1x…"
          },
          "url": {
            "type": "string",
            "example": "https://ops.example.com/hooks/rivora"
          },
          "events": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "revenue.received",
                "revenue.settled",
                "risk.assessment.completed",
                "credit.limit.updated",
                "credit.draw.completed",
                "credit.repayment.completed",
                "borrower.watchlisted",
                "borrower.restricted",
                "borrower.defaulted",
                "vault.utilization.changed"
              ]
            }
          },
          "secretFingerprint": {
            "type": "string",
            "example": "9f2c11ab",
            "description": "First bytes of the secret’s SHA-256 — enough to tell two secrets apart, never enough to sign."
          },
          "active": {
            "type": "boolean",
            "example": true
          },
          "createdAt": {
            "type": "string",
            "example": "2026-08-05T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "url",
          "events",
          "secretFingerprint",
          "active",
          "createdAt"
        ]
      },
...
    }
  }
}